refactor: share release pipeline options
