+++ src/core/release/execution_dispatch.rs
@@ -75,9 +75,10 @@ pub(super) fn execute_release_plan_step(
                 .and_then(|value| value.as_str())
                 .unwrap_or_default();
             Ok(Some(
-                executor::run_artifact_inventory(&mut context.state, dir).unwrap_or_else(|err| {
-                    failed_result("artifacts.inventory", "artifacts.inventory", err)
-                }),
+                executor::artifacts::run_artifact_inventory(&mut context.state, dir)
+                    .unwrap_or_else(|err| {
+                        failed_result("artifacts.inventory", "artifacts.inventory", err)
+                    }),
             ))
         }
         "git.tag" => {
