+++ src/core/release/executor.rs
@@ -22,6 +22,7 @@ use crate::core::release::{changelog as release_changelog, version};
 use super::types::{ReleaseArtifact, ReleaseState, ReleaseStepResult, ReleaseStepStatus};
 use super::utils::{extract_latest_notes, parse_release_artifacts};
 
+pub(crate) mod artifacts;
 pub(crate) mod changelog;
 pub(crate) mod prepare;
 
@@ -585,86 +586,6 @@ pub(crate) fn run_package(
     });
 
     Ok(step_success("package", "package", Some(data), Vec::new()))
-}
-
-/// Inventory artifacts that were already built by an external release build.
-/// This lets `homeboy release --head --from-artifacts <dir>` reuse the normal
-/// github.release and publish steps without re-running release.package.
-pub(crate) fn run_artifact_inventory(
-    state: &mut ReleaseState,
-    artifact_dir: &str,
-) -> Result<ReleaseStepResult> {
-    let dir = std::path::Path::new(artifact_dir);
-    if !dir.is_dir() {
-        return Err(Error::validation_invalid_argument(
-            "from-artifacts",
-            format!("Artifact directory '{}' does not exist", artifact_dir),
-            Some(artifact_dir.to_string()),
-            None,
-        ));
-    }
-
-    let mut artifacts = Vec::new();
-    for entry in std::fs::read_dir(dir).map_err(|e| {
-        Error::internal_io(
-            format!(
-                "Failed to read artifact directory '{}': {}",
-                artifact_dir, e
-            ),
-            Some(artifact_dir.to_string()),
-        )
-    })? {
-        let entry = entry.map_err(|e| {
-            Error::internal_io(
-                format!("Failed to read artifact directory entry: {}", e),
-                Some(artifact_dir.to_string()),
-            )
-        })?;
-        let path = entry.path();
-        if !path.is_file() {
-            continue;
-        }
-        let canonical = std::fs::canonicalize(&path).map_err(|e| {
-            Error::internal_io(
-                format!(
-                    "Failed to resolve artifact path '{}': {}",
-                    path.display(),
-                    e
-                ),
-                Some(path.display().to_string()),
-            )
-        })?;
-        artifacts.push(ReleaseArtifact {
-            path: canonical.display().to_string(),
-            artifact_type: None,
-            platform: None,
-        });
-    }
-
-    artifacts.sort_by(|a, b| a.path.cmp(&b.path));
-    if artifacts.is_empty() {
-        return Err(Error::validation_invalid_argument(
-            "from-artifacts",
-            format!("Artifact directory '{}' contains no files", artifact_dir),
-            Some(artifact_dir.to_string()),
-            None,
-        ));
-    }
-
-    let artifact_count = artifacts.len();
-    state.artifacts.extend(artifacts.clone());
-    let data = serde_json::json!({
-        "dir": artifact_dir,
-        "artifact_count": artifact_count,
-        "artifacts": artifacts,
-    });
-
-    Ok(step_success(
-        "artifacts.inventory",
-        "artifacts.inventory",
-        Some(data),
-        Vec::new(),
-    ))
 }
 
 /// Invoke the `release.publish` action on the named extension.
@@ -1344,8 +1265,8 @@ fn sanitize_tag_for_filename(tag: &str) -> String {
 #[cfg(test)]
 mod tests {
     use super::{
-        fallback_gh_command, publish_step_result, run_artifact_inventory, run_git_push,
-        sanitize_tag_for_filename, store_artifacts_from_output,
+        fallback_gh_command, publish_step_result, run_git_push, sanitize_tag_for_filename,
+        store_artifacts_from_output,
     };
     use crate::core::component::Component;
     use crate::core::release::ReleaseStepStatus;
@@ -1399,28 +1320,6 @@ mod tests {
                 .data
                 .and_then(|data| data.get("success").and_then(serde_json::Value::as_bool)),
             Some(false)
-        );
-    }
-
-    #[test]
-    fn artifact_inventory_loads_existing_files_into_release_state() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let artifact_path = temp.path().join("homeboy.tar.gz");
-        std::fs::write(&artifact_path, "artifact").expect("write artifact");
-        std::fs::create_dir(temp.path().join("nested")).expect("nested dir");
-
-        let mut state = crate::core::release::types::ReleaseState::default();
-        let result = run_artifact_inventory(&mut state, &temp.path().to_string_lossy())
-            .expect("inventory should succeed");
-
-        assert_eq!(result.status, ReleaseStepStatus::Success);
-        assert_eq!(state.artifacts.len(), 1);
-        assert_eq!(
-            state.artifacts[0].path,
-            std::fs::canonicalize(&artifact_path)
-                .expect("canonical artifact")
-                .display()
-                .to_string()
         );
     }
