+++ src/core/release/mod.rs
@@ -26,8 +26,8 @@ pub use planner::plan;
 pub use types::{
     BatchReleaseComponentResult, BatchReleaseResult, BatchReleaseSummary, ReleaseArtifact,
     ReleaseCommandInput, ReleaseCommandResult, ReleaseDeploymentResult, ReleaseDeploymentSummary,
-    ReleaseOptions, ReleasePlan, ReleaseProjectDeployResult, ReleaseRun, ReleaseRunResult,
-    ReleaseRunSummary, ReleaseStepResult, ReleaseStepStatus,
+    ReleaseOptions, ReleasePipelineOptions, ReleasePlan, ReleaseProjectDeployResult, ReleaseRun,
+    ReleaseRunResult, ReleaseRunSummary, ReleaseStepResult, ReleaseStepStatus,
 };
 pub use utils::{extract_latest_notes, parse_release_artifacts};
 pub use workflow::{run_batch, run_command};
