+++ src/core/release/types.rs
@@ -201,15 +201,7 @@ pub struct ReleaseState {
 }
 
 #[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct ReleaseOptions {
-    pub bump_type: String,
-    pub dry_run: bool,
-    /// Override the component's `local_path` for this release.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub path_override: Option<String>,
-    /// Skip lint/test code quality checks before release.
-    #[serde(default)]
-    pub skip_checks: bool,
+pub struct ReleasePipelineOptions {
     /// Skip publish/package steps (version bump + tag + push only).
     /// Use when CI handles publishing after the tag is pushed.
     #[serde(default)]
@@ -223,6 +215,20 @@ pub struct ReleaseOptions {
     /// Deploy after release — defers artifact cleanup until after deployment.
     #[serde(default)]
     pub deploy: bool,
+}
+
+#[derive(Debug, Clone, Default, Serialize, Deserialize)]
+pub struct ReleaseOptions {
+    pub bump_type: String,
+    pub dry_run: bool,
+    /// Override the component's `local_path` for this release.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub path_override: Option<String>,
+    /// Skip lint/test code quality checks before release.
+    #[serde(default)]
+    pub skip_checks: bool,
+    #[serde(default, flatten)]
+    pub pipeline: ReleasePipelineOptions,
     /// Skip the GitHub Release creation step (tag + notes on github.com).
     /// Use when another pipeline (CI, semantic-release, etc.) already owns that step.
     #[serde(default)]
@@ -262,13 +268,7 @@ pub struct ReleaseCommandInput {
     #[serde(default)]
     pub dry_run: bool,
     #[serde(default)]
-    pub deploy: bool,
-    #[serde(default)]
     pub recover: bool,
-    #[serde(default)]
-    pub head: bool,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub from_artifacts: Option<String>,
     #[serde(default)]
     pub skip_checks: bool,
     /// Explicit bump override: "major", "minor", "patch", or a version string like "2.0.0".
@@ -278,8 +278,8 @@ pub struct ReleaseCommandInput {
     /// Permit a keyword bump lower than the commit-derived recommendation.
     #[serde(default)]
     pub force_lower_bump: bool,
-    #[serde(default)]
-    pub skip_publish: bool,
+    #[serde(default, flatten)]
+    pub pipeline: ReleasePipelineOptions,
     /// Skip the GitHub Release creation step (tag + notes on github.com).
     #[serde(default)]
     pub skip_github_release: bool,
