+++ src/core/release/workflow.rs
@@ -13,16 +13,16 @@ pub fn run_command(input: ReleaseCommandInput) -> Result<(ReleaseCommandResult, 
         return run_recover(&input);
     }
 
-    if input.from_artifacts.is_some() && !input.head {
+    if input.pipeline.from_artifacts.is_some() && !input.pipeline.head {
         return Err(Error::validation_invalid_argument(
             "from-artifacts",
             "--from-artifacts requires --head",
-            input.from_artifacts.clone(),
+            input.pipeline.from_artifacts.clone(),
             None,
         ));
     }
 
-    if input.head && input.bump_override.is_some() {
+    if input.pipeline.head && input.bump_override.is_some() {
         return Err(Error::validation_invalid_argument(
             "bump",
             "--head uses the version already present at HEAD and cannot be combined with --bump",
@@ -40,7 +40,7 @@ pub fn run_command(input: ReleaseCommandInput) -> Result<(ReleaseCommandResult, 
     )?;
 
     let monorepo = git::MonorepoContext::detect(&component.local_path, &input.component_id);
-    let resolved_bump = if input.head {
+    let resolved_bump = if input.pipeline.head {
         None
     } else {
         resolve_bump(&component.local_path, monorepo.as_ref())?
@@ -52,7 +52,7 @@ pub fn run_command(input: ReleaseCommandInput) -> Result<(ReleaseCommandResult, 
     let has_breaking_commits = auto_bump_type == "major";
 
     // Resolve the effective bump type: --bump overrides auto-detection.
-    let bump_type = if input.head {
+    let bump_type = if input.pipeline.head {
         "head".to_string()
     } else if let Some(ref override_value) = input.bump_override {
         // Check if it's an explicit version string (e.g. "2.0.0")
@@ -138,10 +138,7 @@ pub fn run_command(input: ReleaseCommandInput) -> Result<(ReleaseCommandResult, 
         dry_run: input.dry_run,
         path_override: input.path_override,
         skip_checks: input.skip_checks,
-        skip_publish: input.skip_publish,
-        head: input.head,
-        from_artifacts: input.from_artifacts.clone(),
-        deploy: input.deploy,
+        pipeline: input.pipeline.clone(),
         skip_github_release: input.skip_github_release,
         git_identity: input.git_identity.clone(),
         bump_policy: ReleaseBumpPolicyOptions {
@@ -153,7 +150,7 @@ pub fn run_command(input: ReleaseCommandInput) -> Result<(ReleaseCommandResult, 
 
     if options.dry_run {
         let plan = super::plan(&input.component_id, &options)?;
-        let new_version = if input.head {
+        let new_version = if input.pipeline.head {
             current_component_version(&component)?
         } else {
             extract_new_version_from_plan(&plan)
@@ -162,6 +159,7 @@ pub fn run_command(input: ReleaseCommandInput) -> Result<(ReleaseCommandResult, 
             .as_ref()
             .map(|v| format_tag(v, monorepo.as_ref()));
         let deployment = input
+            .pipeline
             .deploy
             .then(|| super::deployment::plan_deployment(&input.component_id));
 
@@ -185,7 +183,7 @@ pub fn run_command(input: ReleaseCommandInput) -> Result<(ReleaseCommandResult, 
     let (plan, run_result) = super::pipeline::run_with_plan(&input.component_id, &options)?;
     display_release_summary(&run_result);
 
-    let new_version = if input.head {
+    let new_version = if input.pipeline.head {
         current_component_version(&component)?
     } else {
         extract_new_version_from_run(&run_result)
@@ -609,14 +607,11 @@ pub fn run_batch(
             component_id: component_id.clone(),
             path_override: None,
             dry_run: input_template.dry_run,
-            deploy: input_template.deploy,
             recover: input_template.recover,
-            head: input_template.head,
-            from_artifacts: input_template.from_artifacts.clone(),
             skip_checks: input_template.skip_checks,
             bump_override: input_template.bump_override.clone(),
             force_lower_bump: input_template.force_lower_bump,
-            skip_publish: input_template.skip_publish,
+            pipeline: input_template.pipeline.clone(),
             skip_github_release: input_template.skip_github_release,
             git_identity: input_template.git_identity.clone(),
         };
